+++ Options/FireHelicopterDepotOptions.cs
@@ -17,7 +17,7 @@ public class FireHelicopterDepotOptions :
     public int NoisePollution { get; set; }
     public int FireEngineCapacity { get; set; }
     public int FireHelicopterCapacity { get; set; }
-    public int VehicleEfficiency { get; set; }
+    public float VehicleEfficiency { get; set; }
     public int DisasterResponseCapacity { get; set; }
 
 }
